Get rid of overengineered macro
While the macro's implementation was fine based on the documentation for CMake 3.16, which we (currently) use, later versions of CMake recommend against this approach.

https://gitlab.com/OpenMW/openmw/-/merge_requests/4777 has an implementation that picks between a native feature added later that does the same job if it's available, but uses the one we already had if it isn't.
That's not ideal as the implementation ends up complicated and we'd have two code paths to support.

There's a simpler approach where we just temporarily set the directory property and call the real function, which would probably be the best approach in the general case.

However, as we want to set the directory property for everything in the directory, we can just set it on the `extern` directory, and then we don't need the macro at all.