Don't mix line endings
This doesn't fix the BOM in the middle of the merged CI-ID.txt, but that shouldn't be happening in the first place.
GitLab's docs say the default is pwsh rather than powershell (i.e. PowerShell >= 6, not 5.1), which defaults to non-BOM UTF-8, so there should be no BOM to contend with.