Balance build times between jobs
Before this, there was a difference of about nineteen minutes.

All the things that aren't the engine (except its dependencies) that were in group one take less than nineteen minutes to build in total, so moving everything isn't enough to give total balance.

openmw-tests depends on openmw-lib, so you have to build basically the whole engine to build it, so it should stay in the same group.